bugfix collision

+++ lib/my/include/my.h
@@ -27,7 +27,7 @@ int my_printf(char *str, ...);
 int my_putlongnbr(long int nb);
 int my_putlonglongnbr(long long int nb);
 int my_strncmp(char const *s1, char const *s2, int n);
-int dist(int xA, int yA, int xB, int yB);
+double dist(double xA, double yA, double xB, double yB);
 int my_strcmp(char const *s1, char const *s2);
